intermediate changes for team sync
